[SW-47] refactor(season & team): 엔드포인트 수정
