+++ src/main/java/com/github/kmu_shell_we/domain/season/_team/_user_team/controller/AdminUserTeamController.java
@@ -1,16 +1,16 @@
-package com.github.kmu_shell_we.domain.team._userTeam.controller;
+package com.github.kmu_shell_we.domain.season._team._user_team.controller;
 
-import com.github.kmu_shell_we.domain.team._userTeam.service.AdminUserTeamService;
+import com.github.kmu_shell_we.domain.season._team._user_team.service.AdminUserTeamService;
 import com.github.kmu_shell_we.global.response.ApiResponse;
+import com.github.kmu_shell_we.global.security.guard.AdminGuard;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.UUID;
 
-@PreAuthorize("hasRole('ADMIN')")
+@AdminGuard
 @RestController
 @RequestMapping("/admin/seasons/{seasonId}/teams/{teamId}/members/{userId}")
 @RequiredArgsConstructor
