+++ src/main/java/com/github/kmu_shell_we/domain/season/exception/SeasonExceptions.java
@@ -6,9 +6,9 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum SeasonExceptionCode implements ApiExceptionCode {
+public enum SeasonExceptions implements ApiExceptionCode {
 
-    NOT_FOUND("SEASON_001", "시즌을 찾을 수 없습니다."),
+    NOT_FOUND_SEASON("SEASON_001", "시즌을 찾을 수 없습니다."),
     NOT_FOUND_CURRENT_SEASON("SEASON_002", "진행중인 시즌을 찾을 수 없습니다."),
     ;
